refactor: 把 hub 迁到 @hmr/host-runtime，host 不再留任何插件
hub / resolveCatalog 随 host-runtime 导出；cordis.plugins.json 的 host
表最后一项加载它。host/index.ts 只读配置并 plugin()。
