+++ docs/plugin-packages.md
@@ -2,15 +2,15 @@
 
 主仓 **`host/` / `src/` 只留加载器**：读 `cordis.plugins.json`，按包名动态 import。插件源码在 `packages/`。
 
+`host/index.ts` 只做：建 Context、读 `host` 表、`plugin()`。`hub` 也是 `@hmr/host-runtime/hub`，写在 `host` 表最后一项（它再去挂 `plugins` 表）。
+
 ## 三张表
 
 | 键 | 谁加载 | 含义 |
 |---|---|---|
 | `host` | `host/index.ts` | 内核（http / sessions / tools …），按数组顺序 `await plugin` |
 | `web` | `src/main.tsx` ← `virtual:cordis-web-runtime` | 壳（slots / shell / ui-hub …） |
-| `plugins` | host `hub` + web `ui-hub` | 可热插拔能力；`ui` 字段成对卸前端 |
-
-例外：`hub` 本身仍在 `host/plugins/registry/hub.ts`，因为它要 `resolveCatalog()` 去挂 `plugins`。
+| `plugins` | `@hmr/host-runtime/hub` + web `ui-hub` | 可热插拔能力；`ui` 字段成对卸前端 |
 
 ## 包
