+++ packages/host-runtime/src/registry/resolve-catalog.test.ts
@@ -1,12 +1,12 @@
 import { test } from 'vitest'
 import assert from 'node:assert/strict'
 import { readFile } from 'node:fs/promises'
-import { join, dirname } from 'node:path'
-import { fileURLToPath } from 'node:url'
+import { join } from 'node:path'
 import { resolveCatalog } from './resolve-catalog.ts'
+import { rootDirFrom } from '../../../../host/cordis-plugins.ts'
 
 test('resolveCatalog loads only packages declared in cordis.plugins.json', async () => {
-  const root = join(dirname(fileURLToPath(import.meta.url)), '..')
+  const root = rootDirFrom()
   const raw = JSON.parse(await readFile(join(root, 'cordis.plugins.json'), 'utf8')) as {
     plugins: Array<{ id: string; package: string; ui?: string }>
   }
